+++ boot/src/memory.c
@@ -2,7 +2,7 @@
 #include <boot/include/bios.h>
 #include <boot/include/multiboot.h>
 
-uintptr_t Memory_StoreBasicInfo(uintptr_t addr) {
+uint32_t Memory_StoreBasicInfo(uint32_t addr) {
 
 	uint32_t* mem = (uint32_t*)(addr);
 	
@@ -45,7 +45,7 @@ uintptr_t Memory_StoreBasicInfo(uintptr_t addr) {
 	return addr + 8;
 }
 
-uintptr_t Memory_StoreE820Info(uintptr_t addr) {
+uint32_t Memory_StoreE820Info(uint32_t addr) {
 
 	// INT 0x15, EAX = 0xE820 is the best way to get the detailed memory map of the system
 	// Memory map is set up at the designated memory address in ES:DI
@@ -92,21 +92,21 @@ uintptr_t Memory_StoreE820Info(uintptr_t addr) {
 }
 
 
-uint64_t Memory_MaxPhysicalAddress(struct Multiboot_E820_Entry* E820_Table, size_t E820_Table_size) {
+uint64_t Memory_MaxPhysicalAddress(struct Multiboot_E820_Entry* E820_Table, uint32_t E820_Table_size) {
 
 	// Find the maximum memory address that exists physically and is available
 	// Iterate over the E820 memory map to get this information
 
 	uint64_t mem_max = 0;
 	if (E820_Table_size == 0 || E820_Table == MEMORY_NULL_PTR) return mem_max;
-	for (size_t i = 0; i < E820_Table_size; i++) {
+	for (uint32_t i = 0; i < E820_Table_size; i++) {
 		if (E820_Table[i].acpi3 != MULTIBOOT_MEMORY_ACPI3_FLAG || E820_Table[i].type != MULTIBOOT_MEMORY_AVAILABLE) continue;
 		if (E820_Table[i].base + E820_Table[i].size > mem_max) mem_max = E820_Table[i].base + E820_Table[i].size;
 	}
 	return mem_max;
 }
 
-bool Memory_IsPresent(uint64_t min, uint64_t max, struct Multiboot_E820_Entry* E820_Table, size_t E820_Table_size) {
+bool Memory_IsPresent(uint64_t min, uint64_t max, struct Multiboot_E820_Entry* E820_Table, uint32_t E820_Table_size) {
 
 	if (min >= max) return false;
 	if (E820_Table_size == 0 || E820_Table == MEMORY_NULL_PTR) return false;
@@ -115,13 +115,13 @@ bool Memory_IsPresent(uint64_t min, uint64_t max, struct Multiboot_E820_Entry* E
 	if (min >= mem_max || max > mem_max) return false;
 	
 	// First check if the memory range is covered by a single entry 
-	for (size_t i = 0; i < E820_Table_size; i++) {
+	for (uint32_t i = 0; i < E820_Table_size; i++) {
 		if (E820_Table[i].acpi3 != MULTIBOOT_MEMORY_ACPI3_FLAG || E820_Table[i].type != MULTIBOOT_MEMORY_AVAILABLE) continue;
 		if (E820_Table[i].base <= min && E820_Table[i].base + E820_Table[i].size >= max) return true;
 	}
 	
 	// Next check if the memory range is covered by multiple entries
-	for (size_t i = 0; i < E820_Table_size; i++) {
+	for (uint32_t i = 0; i < E820_Table_size; i++) {
 		uint64_t e820_entry_end = E820_Table[i].base + E820_Table[i].size;
 		if (E820_Table[i].acpi3 != MULTIBOOT_MEMORY_ACPI3_FLAG || E820_Table[i].type != MULTIBOOT_MEMORY_AVAILABLE) continue;
 		if (E820_Table[i].base <= min && e820_entry_end > min) return Memory_IsPresent(e820_entry_end, max, E820_Table, E820_Table_size);
@@ -130,8 +130,8 @@ bool Memory_IsPresent(uint64_t min, uint64_t max, struct Multiboot_E820_Entry* E
 	return false;
 }
 
-uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E820_Entry* E820_Table, size_t E820_Table_size) {
-	size_t table_size  = 0;
+uint32_t Memory_StoreInfo(uint32_t addr, bool page_align, struct Multiboot_E820_Entry* E820_Table, uint32_t E820_Table_size) {
+	uint32_t table_size  = 0;
 	struct Boot_Block64* table_ptr = (struct Boot_Block64*)(addr);
 	table_ptr->address = 0;
 	table_ptr->size = 0;
@@ -141,7 +141,7 @@ uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E82
 	while (table_ptr->address < mem_end) {
 		uint64_t addr_limit = ~((uint64_t)0);
 		uint64_t closest_boundary = addr_limit;
-		for (size_t i = 0; i < E820_Table_size; i++) {
+		for (uint32_t i = 0; i < E820_Table_size; i++) {
 			if (E820_Table[i].acpi3 != MULTIBOOT_MEMORY_ACPI3_FLAG || E820_Table[i].type != MULTIBOOT_MEMORY_AVAILABLE) continue;
 			if (E820_Table[i].base >= table_ptr->address && E820_Table[i].base < closest_boundary) closest_boundary = E820_Table[i].base;
 		}
@@ -149,7 +149,7 @@ uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E82
 		
 		table_ptr->address = closest_boundary;
 		table_ptr->size = 0;
-		for (size_t i = 0; i < E820_Table_size; i++) {
+		for (uint32_t i = 0; i < E820_Table_size; i++) {
 			if (E820_Table[i].acpi3 != MULTIBOOT_MEMORY_ACPI3_FLAG || E820_Table[i].type != MULTIBOOT_MEMORY_AVAILABLE) continue;
 			uint64_t e820_entry_end = E820_Table[i].base + E820_Table[i].size;
 			if (E820_Table[i].base <= table_ptr->address && e820_entry_end >= table_ptr->address + table_ptr->size) table_ptr->size = e820_entry_end - table_ptr->address;
@@ -163,7 +163,7 @@ uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E82
 	
 	// Adjust the entries on 4KB boundaries
 	if (page_align) {
-		for (size_t i = 0; i < table_size; i++) {
+		for (uint32_t i = 0; i < table_size; i++) {
 			uint64_t min_addr = table_ptr[i].address;
 			uint64_t mask = ~((uint64_t)(0xFFF));
 			if ((min_addr & mask) < min_addr) min_addr = (min_addr & mask) + 0x1000;
@@ -180,9 +180,9 @@ uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E82
 	}
 	
 	// Remove entries that are reduced to 0-size after 4KB boundary adjustment
-	for (size_t i = 0; i < table_size; i++) {
+	for (uint32_t i = 0; i < table_size; i++) {
 		if (table_ptr[i].size == 0) {
-			for (size_t j = i+i; j < table_size; j++) {
+			for (uint32_t j = i+i; j < table_size; j++) {
 				table_ptr[j-1].address = table_ptr[j].address;
 				table_ptr[j-1].size = table_ptr[j].size;
 			}
@@ -193,7 +193,7 @@ uintptr_t Memory_StoreInfo(uintptr_t addr, bool page_align, struct Multiboot_E82
 	return addr + table_size * sizeof(struct Boot_Block64);
 }
 
-uint32_t Memory_BlockAboveAddress(uint32_t addr, uint32_t size, uint32_t align, struct Boot_Block64* mmap, size_t mmap_size) {
+uint32_t Memory_BlockAboveAddress(uint32_t addr, uint32_t size, uint32_t align, struct Boot_Block64* mmap, uint32_t mmap_size) {
 
     if (addr == MEMORY_32BIT_LIMIT || addr + size < addr) return MEMORY_32BIT_LIMIT;
    
@@ -203,7 +203,7 @@ uint32_t Memory_BlockAboveAddress(uint32_t addr, uint32_t size, uint32_t align, 
     uint32_t aligned_addr = (addr % align == 0 || align <= 1 ? addr : align * (1 + addr/align));
    
     if (mmap_size == 0 || mmap == MEMORY_NULL_PTR) return mem;
-    for (size_t i = 0; i < mmap_size; i++) {
+    for (uint32_t i = 0; i < mmap_size; i++) {
         if (mmap[i].address >= MEMORY_32BIT_LIMIT) continue;
 
         uint32_t mmap_entry_base = (uint32_t)mmap[i].address;
@@ -220,13 +220,13 @@ uint32_t Memory_BlockAboveAddress(uint32_t addr, uint32_t size, uint32_t align, 
     return mem;
 }
 
-uint32_t Memory_BlockBelowAddress(uint32_t addr, uint32_t size, uint32_t align, struct Boot_Block64* mmap, size_t mmap_size) {
+uint32_t Memory_BlockBelowAddress(uint32_t addr, uint32_t size, uint32_t align, struct Boot_Block64* mmap, uint32_t mmap_size) {
 
     uint32_t mem = 0;
     uint32_t aligned_addr = (addr % align == 0 || align <= 1 ? addr - size + 1 : align * ((addr - size + 1)/align));
 
     if (mmap_size == 0 || mmap == MEMORY_NULL_PTR) return mem;
-    for (size_t i = 0; i < mmap_size; i++) {
+    for (uint32_t i = 0; i < mmap_size; i++) {
         if (mmap[i].address >= MEMORY_32BIT_LIMIT) continue;
 
         uint32_t mmap_entry_base = (uint32_t)mmap[i].address;
